relocate get request to feed multiple child
