+++ src/components/pages/AccountDashboard/AccountDashboard.js
@@ -1,25 +1,8 @@
-import React, {useState, useEffect} from "react";
+import React from "react";
 import Avatar from "./worker.png";
-import AuthService from "../../../Services/AuthService";
 import "./styles/AccountDashboard.css";
-function AccountDashboard() {
-  const [user, setUser] = useState({});
-  const [project, setProject] = useState(null);
-  const getData = async () => {
-    try {
-      const {data} = await AuthService.getInfo("5f5413bc90d4c22f502d97ec");
-      setUser(data)
-      setProject(data.project.length);
-    }
-    catch (err) {
-      console.log(`Error - AccountDashboard.js - ${err}`);
-    }
-  }
-  
-  useEffect(() => {
-    getData({});
-  }, []);
-  
+function AccountDashboard(props) {
+
   return (
       <div className="profile">
         <div className="profile__picture">
@@ -27,7 +10,7 @@ function AccountDashboard() {
         </div>
         <div className="profile__header">
           <div className="profile__account">
-            <h4 className="profile__username">{user.firstName}{" "}{user.lastName}</h4>
+            <h4 className="profile__username">{props.username}</h4>
           </div>
         </div>
         <div className="profile__stats">
@@ -36,7 +19,7 @@ function AccountDashboard() {
               <i className="material-icons">engineering</i>
             </div>
             <div className="profile__value">
-              {project}
+              {props.project}
               <div className="profile__key">Projects</div>
             </div>
           </div>
